added code for reading onboard processor temperature using xadc
also added a small python script for converting the serial output into
something usable, plotted it in matlab, and i think theres an issue
with one of the temperature conversion scaling factors.

temperature output in the serial console varies with applied heat, but
cannot get a solid physical value from it yet. so far so good though.

The included .bit file outputs the temperature value to the serial
stream, along with initializes the via display(should be all blue)