fixed a couple issue with the pwm driver
clock divider was acting strange so I changed the numbers based on
measured frequencies(tried to obtain a 300 to 500Hz signal). Seems to
be okay now.

Tries using the IR510 power nfet for 3.3V logic switching to drive the
fan motor, but turns out the Vgs(th) value of this is 4V max, so the
FPGA can’t switch it. Should have checked before I bought two of these…
Currently looking at the IRLB80721 nfet with a Vgs(th) of 2.35V max as
the potential replacement.